+++ LogicBuilder.EntityFrameworkCore.Tests/Data/CourseAssignment.cs
@@ -7,7 +7,7 @@ namespace LogicBuilder.EntityFrameworkCore.Tests.Data
     {
         public int InstructorID { get; set; }
         public int CourseID { get; set; }
-        public Instructor Instructor { get; set; }
-        public Course Course { get; set; }
+        public Instructor? Instructor { get; set; }
+        public Course? Course { get; set; }
     }
 }