+++ LogicBuilder.EntityFrameworkCore.Tests/ExpressionStringBuilder.cs
@@ -47,7 +47,7 @@ namespace LogicBuilder.EntityFrameworkCore.Tests
 
         protected override Expression VisitParameter(ParameterExpression node)
         {
-            Out(node.Name);
+            Out(node.Name!);
             return node;
         }
 
@@ -57,9 +57,9 @@ namespace LogicBuilder.EntityFrameworkCore.Tests
             if (node.Expression == null && node.NodeType == ExpressionType.MemberAccess)
             {
                 Visit(node.Expression);
-                Out(node.Member.DeclaringType.Name + "." + node.Member.Name);
+                Out(node.Member.DeclaringType!.Name + "." + node.Member.Name);
             }
-            else if (node.Expression.NodeType == ExpressionType.Constant)
+            else if (node.Expression?.NodeType == ExpressionType.Constant)
             {
                 Visit(node.Expression);
                 if (!typeof(ConstantContainer).IsAssignableFrom(node.Expression?.Type))
@@ -84,13 +84,13 @@ namespace LogicBuilder.EntityFrameworkCore.Tests
             {
                 Out
                 (
-                    GetOutString(GetConstantValue())
+                    GetOutString(GetConstantValue() ?? "")
                 );
             }
 
             return node;
 
-            object GetConstantValue()
+            object? GetConstantValue()
                 => node.Value is ConstantContainer constantContainer
                     ? constantContainer.Property
                     : node.Value;